arfile.ArFile.extractfile: Fix an incorrect early return.
Now it will return a matching member even if it is not the first one in the
archive.  This commit also adds a TODO to investigate either switching that
method to call getmember instead of looping over the memembers, or
deprecating/deleting it if nobody is actually using it.

Thanks to Mika Eloranta for the bug report and patch.

Closes: #689313